Move more functions to hooks
