+++ front/components/assistant/conversation/ConversationMenu.tsx
@@ -19,12 +19,12 @@ import { useCallback, useEffect, useState } from "react";
 import { DeleteConversationsDialog } from "@app/components/assistant/conversation/DeleteConversationsDialog";
 import { EditConversationTitleDialog } from "@app/components/assistant/conversation/EditConversationTitleDialog";
 import { LeaveConversationDialog } from "@app/components/assistant/conversation/LeaveConversationDialog";
+import { useDeleteConversation } from "@app/hooks/useDeleteConversation";
 import { useSendNotification } from "@app/hooks/useNotification";
 import { useURLSheet } from "@app/hooks/useURLSheet";
 import {
   useConversationParticipants,
   useConversationParticipationOption,
-  useDeleteConversation,
   useJoinConversation,
 } from "@app/lib/swr/conversations";
 import { useUser } from "@app/lib/swr/user";
